slack_bot/requirements.txt
psycopg2-binary
sqlalchemy
pandas
slackclient==2.0.1
config
